requirements.txt
# Core Framework
flask>=2.3.0,<4.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0

# LangChain Core (v0.3 is current stable)
langchain>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0

# LangChain Integrations
langchain-openai>=0.2.0,<0.4.0
langchain-anthropic>=0.2.0,<0.4.0
langchain-google-genai>=2.0.0,<3.0.0
langchain-pinecone>=0.2.0,<0.3.0
langchain-huggingface>=0.1.0,<0.2.0

# Document Processing
pypdf>=4.0.0
sentence-transformers>=2.2.2

# Vector Database
pinecone-client>=4.0.0

# Security & Rate Limiting
flask-limiter>=4.0.0
flask-talisman>=1.1.0
bleach>=6.1.0
cryptography>=41.0.0

# Advanced Features
redis>=5.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Monitoring & Logging
prometheus-flask-exporter>=0.23.0

# Development & Testing
pytest>=7.4.0
pytest-flask>=1.3.0
black>=23.0.0
flake8>=6.0.0
